add item click listener
