+++ app/src/main/java/ru/yandex/yamblz/ui/adapter/DeveloperSettingsSpinnerAdapter.java
@@ -1,4 +1,4 @@
-package ru.yandex.yamblz.ui.adapters;
+package ru.yandex.yamblz.ui.adapter;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
