--- app/src/main/java/ru/yandex/yamblz/ui/fragments/ContentAdapter.java
@@ -1,116 +1,0 @@
-package ru.yandex.yamblz.ui.fragments;
-
-import android.graphics.Color;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Random;
-
-import ru.yandex.yamblz.R;
-
-class ContentAdapter extends RecyclerView.Adapter<ContentAdapter.ContentHolder> implements ItemTouchHelperAdapter {
-
-    private final Random rnd = new Random();
-    private final List<Integer> colors = new ArrayList<>();
-
-    @Override
-    public ContentHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        return new ContentHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.content_item, parent, false));
-    }
-
-    @Override
-    public void onBindViewHolder(ContentHolder holder, int position) {
-        holder.bind(createColorForPosition(position));
-    }
-
-    @Override
-    public int getItemCount() {
-        return Integer.MAX_VALUE;
-    }
-
-    private Integer createColorForPosition(int position) {
-        if (position >= colors.size()) {
-            colors.add(Color.rgb(rnd.nextInt(255), rnd.nextInt(255), rnd.nextInt(255)));
-        }
-        return colors.get(position);
-    }
-
-    static class ContentHolder extends RecyclerView.ViewHolder {
-        ContentHolder(View itemView) {
-            super(itemView);
-        }
-
-        void bind(Integer color) {
-            itemView.setBackgroundColor(color);
-            ((TextView) itemView).setText("#".concat(Integer.toHexString(color).substring(2)));
-        }
-    }
-
-    @Override
-    public void onItemDismiss(int position) {
-        colors.remove(position);
-        notifyItemRemoved(position);
-    }
-
-    @Override
-    public void onItemMove(int fromPosition, int toPosition) {
-        if (fromPosition < toPosition) {
-            for (int i = fromPosition; i < toPosition; i++) {
-                Collections.swap(colors, i, i + 1);
-            }
-        } else {
-            for (int i = fromPosition; i > toPosition; i--) {
-                Collections.swap(colors, i, i - 1);
-            }
-        }
-        notifyItemMoved(fromPosition, toPosition);
-    }
-
-    static class SimpleItemTouchHelperCallback extends ItemTouchHelper.Callback {
-
-        private final ItemTouchHelperAdapter mAdapter;
-
-        public SimpleItemTouchHelperCallback(
-                ItemTouchHelperAdapter adapter) {
-            mAdapter = adapter;
-        }
-
-        @Override
-        public int getMovementFlags(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder) {
-            int dragFlags = ItemTouchHelper.UP | ItemTouchHelper.DOWN
-                    | ItemTouchHelper.RIGHT | ItemTouchHelper.END
-                    | ItemTouchHelper.LEFT | ItemTouchHelper.START;
-            int swipeFlags = ItemTouchHelper.RIGHT | ItemTouchHelper.END;
-            return makeMovementFlags(dragFlags, swipeFlags);
-        }
-
-        @Override
-        public boolean onMove(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, RecyclerView.ViewHolder target) {
-            mAdapter.onItemMove(viewHolder.getAdapterPosition(),
-                    target.getAdapterPosition());
-            return true;
-        }
-
-        @Override
-        public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
-            mAdapter.onItemDismiss(viewHolder.getAdapterPosition());
-        }
-
-        @Override
-        public boolean isLongPressDragEnabled() {
-            return true;
-        }
-
-        @Override
-        public boolean isItemViewSwipeEnabled() {
-            return true;
-        }
-    }
-}
